refactor: extract conversions with getConversionsForUnit
