+++ lib/ui/settings/conversions.js
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2017 Teppo Kurki <teppo.kurki@iki.fi>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+var Qty = require('js-quantities');
+
+
+var conversions = {
+  "m/s": {
+    "kn": Qty.swiftConverter("m/s", "kn"),
+    "km/h": Qty.swiftConverter("m/s", "km/h")
+  },
+  "K": {
+    "C": Qty.swiftConverter("tempK", "tempC"),
+    "F": Qty.swiftConverter("tempK", "tempF")
+  },
+  "Hz": {
+    "1/min": function(hz) {
+      return hz * 60;
+    }
+  },
+  "m": {
+    "fathom": Qty.swiftConverter('m', 'fathom'),
+    "feet": Qty.swiftConverter('m', 'foot'),
+    "km": Qty.swiftConverter('m', 'km'),
+    "nm": Qty.swiftConverter('m', 'nmi'),
+  },
+  "Pa": {
+    "kPa": Qty.swiftConverter('pascal', 'kPa'),
+    "bar": Qty.swiftConverter('pascal', 'bar'),
+    "psi": Qty.swiftConverter('pascal', 'psi')
+  },
+  "s": {
+    "minutes": Qty.swiftConverter('s', 'minutes'),
+    "hours": Qty.swiftConverter('s', 'hours'),
+    "days": Qty.swiftConverter('s', 'days')
+  },
+  "rad/s": {
+    "deg/s": Qty.swiftConverter('rad/s', 'deg/s'),
+    "deg/min": Qty.swiftConverter('rad/s', 'deg/min')
+  }
+}
+
+
+export function getConversionsForUnit(unit) {
+  return conversions[unit]
+}
