+++ lib/ui/widgets/universal.js
@@ -1,49 +1,10 @@
 var React = require('react');
 var d3 = require('d3');
 var Bacon = require('baconjs');
-var Qty = require('js-quantities');
 var _ = require('lodash');
 
-var conversions = {
-  "m/s": {
-    "kn": Qty.swiftConverter("m/s", "kn"),
-    "km/h": Qty.swiftConverter("m/s", "km/h")
-  },
-  "K": {
-    "C": Qty.swiftConverter("tempK", "tempC"),
-    "F": Qty.swiftConverter("tempK", "tempF")
-  },
-  "Hz": {
-    "1/min": function(hz) {
-      return hz * 60;
-    }
-  },
-  "m": {
-    "fathom": Qty.swiftConverter('m', 'fathom'),
-    "feet": Qty.swiftConverter('m', 'foot'),
-    "km": Qty.swiftConverter('m', 'km'),
-    "nm": Qty.swiftConverter('m', 'nmi'),
-  },
-  "Pa": {
-    "kPa": Qty.swiftConverter('pascal', 'kPa'),
-    "bar": Qty.swiftConverter('pascal', 'bar'),
-    "psi": Qty.swiftConverter('pascal', 'psi')
-  },
-  "s": {
-    "minutes": Qty.swiftConverter('s', 'minutes'),
-    "hours": Qty.swiftConverter('s', 'hours'),
-    "days": Qty.swiftConverter('s', 'days')
-  },
-  "rad/s": {
-    "deg/s": Qty.swiftConverter('rad/s', 'deg/s'),
-    "deg/min": Qty.swiftConverter('rad/s', 'deg/min')
-  }
-}
+import {getConversionsForUnit} from '../settings/conversions'
 
-
-var getConversionsForUnit = function(unit) {
-  return conversions[unit]
-}
 
 var SettingsPanel = (props) => {
   var analogSettings = (
